FieldDefinition as interface #1053
