+++ commercetools-models/src/main/java/io/sphere/sdk/types/FieldDefinition.java
@@ -1,8 +1,7 @@
 package io.sphere.sdk.types;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import io.sphere.sdk.models.Base;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import io.sphere.sdk.models.LocalizedString;
 import io.sphere.sdk.models.TextInputHint;
 
@@ -11,51 +10,25 @@ import javax.annotation.Nullable;
 /**
  * @see Custom
  */
-public final class FieldDefinition extends Base {
-    private final FieldType type;
-    private final String name;
-    private final LocalizedString label;
-    @JsonProperty("required")
-    private final Boolean required;
-    @Nullable
-    private final TextInputHint inputHint;
-
-    @JsonCreator
-    private FieldDefinition(final FieldType type, final String name, final LocalizedString label, final Boolean required, @Nullable final TextInputHint inputHint) {
-        this.type = type;
-        this.name = name;
-        this.label = label;
-        this.required = required;
-        this.inputHint = inputHint;
-    }
-
-    public static FieldDefinition of(final FieldType type, final String name, final LocalizedString label, final Boolean required) {
+@JsonDeserialize(as = FieldDefinitionImpl.class)
+public interface FieldDefinition {
+    static FieldDefinition of(final FieldType type, final String name, final LocalizedString label, final Boolean required) {
         return of(type, name, label, required, null);
     }
 
-    public static FieldDefinition of(final FieldType type, final String name, final LocalizedString label, final Boolean required, final TextInputHint inputHint) {
-        return new FieldDefinition(type, name, label, required, inputHint);
+    static FieldDefinition of(final FieldType type, final String name, final LocalizedString label, final Boolean required, final TextInputHint inputHint) {
+        return new FieldDefinitionImpl(type, name, label, required, inputHint);
     }
 
-    public FieldType getType() {
-        return type;
-    }
+    FieldType getType();
 
-    public String getName() {
-        return name;
-    }
+    String getName();
 
-    public LocalizedString getLabel() {
-        return label;
-    }
+    LocalizedString getLabel();
 
     @JsonProperty("required")
-    public Boolean isRequired() {
-        return required;
-    }
+    Boolean isRequired();
 
     @Nullable
-    public TextInputHint getInputHint() {
-        return inputHint;
-    }
+    TextInputHint getInputHint();
 }
